refactor(mobile): split use-account.ts god-hook (399 → 92+233+202)
use-account.ts was the last 400-LOC god-hook in the mobile package.
PR #101 added 20 tests covering every branch, making this split safe.

Architecture:
- use-account-profile.ts (233 LOC) — profile fetching, name/email/avatar
  updates, preferences, monthly spend fetch.
- use-account-danger.ts (202 LOC) — delete account (two-step confirm),
  sign out, data export, confirmation alerts.
- use-account.ts (92 LOC) — thin compositor that calls both sub-hooks
  and spreads their results. Same `UseAccountArgs` / `UseAccountResult`
  public API as before, so zero caller changes needed.

Behavior preserved:
- fetchMonthlySpend useEffect retains its userId null-guard.
- executeAccountDeletion stays a private useCallback in the danger
  hook, closed over by beginDeleteAccount + confirmDeleteAccountFromModal.
- No key collisions between profile/danger result shapes, no shadowing.

Full `useAccount` suite: 20 tests pass. Typecheck clean.
